+++ packages/css-color-parser/test/util/serialize.mjs
@@ -1,11 +1,11 @@
 import { serializeRGB, serializeP3, serializeOKLCH, serializeHSL } from '@csstools/css-color-parser';
 
-export function serialize_sRGB_data(x) {
+export function serialize_sRGB_data(x, gamutMapping = true) {
 	if (!x) {
 		return '';
 	}
 
-	return serializeRGB(x).toString();
+	return serializeRGB(x, gamutMapping).toString();
 }
 
 export function serialize_HSL_data(x) {
@@ -16,12 +16,12 @@ export function serialize_HSL_data(x) {
 	return serializeHSL(x).toString();
 }
 
-export function serialize_P3_data(x) {
+export function serialize_P3_data(x, gamutMapping = true) {
 	if (!x) {
 		return '';
 	}
 
-	return serializeP3(x).toString();
+	return serializeP3(x, gamutMapping).toString();
 }
 
 export function serialize_OKLCH_data(x) {